Deploy/requirements.txt
streamlit
scikit-learn
tensorflow-cpu
numpy
pandas